Changed the design of the solution into a Command pattern.  In this paradigm the commands are created during the CSV parsing then executed later against the table object. Several Commands have been created and allow a decoupling between the requests from the user to the model. Several unit tests have been added and modified to support this changes
ADDED: Command Interface and Command class implementations to represent each action that can be performed.
ADDED: Custom exception to capture errors caused in the CSV reading.
MODIFIED: Table object to hold a copy of the robot.  This limits the Table to storing a single robot at a time. This should be fixed later if the requirement changes
ADDED: Table interface to aid in decoupling with the commands.
MODIFIED: CsvParser and renamed it to CommandParser to better reflect what it does and also remove confusion between the CSVParser library.
ADDED: two additional sample inputs that can be used to run the application
ADDED: mockito and powermock to the build to allow for testing of the Main method.
added: several test cases to cover all business logic